Changed the configuration for multi-platform.
